+++ modules/core/dependency/python-ihm/ihm/format_bcif.py
@@ -206,6 +206,7 @@ class BinaryCifReader(ihm.format._Reader):
 
     def read_file(self):
         """Read the file and extract data.
+
            :return: True iff more data blocks are available to be read.
         """
         self._add_category_keys()
@@ -549,7 +550,11 @@ class BinaryCifWriter(ihm.format._Writer):
 
     def write_comment(self, comment):
         """See :meth:`ihm.format.CifWriter.write_comment`.
-           @note BinaryCIF does not support comments, so this is a noop"""
+
+           .. note::
+
+              BinaryCIF does not support comments, so this is a noop.
+        """
         pass
 
     def _encode_data(self, data):